move common logic into functions for reuse
