+++ contracts/satellite/src/state.rs
@@ -1,11 +1,15 @@
-use cosmwasm_std::{Addr, Api, DepsMut, Env, StdResult, Timestamp};
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{
+    Addr, Api, CustomMsg, CustomQuery, DepsMut, Env, MessageInfo, Response, StdResult, Timestamp,
+    WasmMsg,
+};
 use cw_storage_plus::{Item, Map};
 
-use crate::error::ContractError;
 use astro_satellite_package::astroport_governance::astroport::common::OwnershipProposal;
-use astro_satellite_package::UpdateConfigMsg;
+use astro_satellite_package::{InstantiateMsg, UpdateConfigMsg};
 use astroport_ibc::{SIGNAL_OUTAGE_LIMITS, TIMEOUT_LIMITS};
-use cosmwasm_schema::cw_serde;
+
+use crate::error::ContractError;
 
 #[cw_serde]
 pub struct Config {
@@ -82,6 +86,92 @@ impl Config {
         }
 
         Ok(())
+    }
+}
+
+pub fn instantiate_state<Q>(
+    deps: DepsMut<Q>,
+    env: Env,
+    msg: InstantiateMsg,
+) -> Result<(), ContractError>
+where
+    Q: CustomQuery,
+{
+    if !TIMEOUT_LIMITS.contains(&msg.timeout) {
+        return Err(ContractError::TimeoutLimitsError {});
+    }
+
+    if !SIGNAL_OUTAGE_LIMITS.contains(&msg.max_signal_outage) {
+        return Err(ContractError::SignalOutageLimitsError {});
+    }
+
+    CONFIG.save(
+        deps.storage,
+        &Config {
+            owner: deps.api.addr_validate(&msg.owner)?,
+            astro_denom: msg.astro_denom,
+            main_controller_port: format!("wasm.{}", msg.main_controller),
+            main_maker: msg.main_maker,
+            gov_channel: None,
+            transfer_channel: msg.transfer_channel,
+            timeout: msg.timeout,
+            max_signal_outage: msg.max_signal_outage,
+            emergency_owner: deps.api.addr_validate(&msg.emergency_owner)?,
+        },
+    )?;
+
+    Ok(LATEST_HUB_SIGNAL_TIME.save(deps.storage, &env.block.time)?)
+}
+
+pub fn update_config<Q, M>(
+    deps: DepsMut<Q>,
+    info: MessageInfo,
+    env: Env,
+    params: UpdateConfigMsg,
+) -> Result<Response<M>, ContractError>
+where
+    Q: CustomQuery,
+    M: CustomMsg,
+{
+    let mut config = CONFIG.load(deps.storage)?;
+    if !(info.sender == config.owner
+        || LATEST_HUB_SIGNAL_TIME
+            .load(deps.storage)?
+            .plus_seconds(config.max_signal_outage)
+            < env.block.time
+            && info.sender == config.emergency_owner)
+    {
+        return Err(ContractError::Unauthorized {});
+    }
+    config.update(deps.api, params)?;
+
+    CONFIG.save(deps.storage, &config)?;
+
+    Ok(Response::new().add_attribute("action", "update_config"))
+}
+
+pub fn set_emegrency_owner_as_admin<Q, M>(
+    deps: DepsMut<Q>,
+    env: Env,
+    info: MessageInfo,
+) -> Result<Response<M>, ContractError>
+where
+    Q: CustomQuery,
+    M: CustomMsg,
+{
+    let config = CONFIG.load(deps.storage)?;
+    if LATEST_HUB_SIGNAL_TIME
+        .load(deps.storage)?
+        .plus_seconds(config.max_signal_outage)
+        < env.block.time
+        && info.sender == config.emergency_owner
+    {
+        Ok(Response::new().add_message(WasmMsg::UpdateAdmin {
+            contract_addr: env.contract.address.to_string(),
+            admin: config.emergency_owner.to_string(),
+        }))
+    } else {
+        Err(ContractError::Unauthorized {})
     }
 }
