+++ contracts/satellite_neutron/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "astro-satellite-neutron"
-version = "1.0.0"
+version = "1.1.0"
 authors = ["Astroport"]
 license = "Apache-2.0"
 description = """
